+++ Orleans.Streams.Kafka/Core/KafkaAdapterReceiver.cs
@@ -5,6 +5,7 @@ using Orleans.Serialization;
 using Orleans.Streams.Kafka.Config;
 using Orleans.Streams.Kafka.Consumer;
 using Orleans.Streams.Utils;
+using Orleans.Streams.Utils.MessageTracking;
 using Orleans.Streams.Utils.Serialization;
 using System;
 using System.Collections.Generic;
@@ -174,13 +175,18 @@ namespace Orleans.Streams.Kafka.Core
 			}
 		}
 
-		private Task TrackMessage(IBatchContainer container)
+		private Task TrackMessage(KafkaBatchContainer container)
 		{
 			if (!_options.MessageTrackingEnabled)
 				return Task.CompletedTask;
 
 			var trackingGrain = _grainFactory.GetMessageTrackerGrain(_queueProperties.QueueName);
-			return trackingGrain.Track(new Immutable<IBatchContainer>(container));
+			return trackingGrain.Track(new Immutable<TrackingUnit>(new TrackingUnit(
+				container.StreamGuid,
+				container.StreamNamespace,
+				container.SequenceToken,
+				container.RawEvents
+			)));
 		}
 	}
 }